+++ zyx-nn/src/embedding.rs
@@ -8,9 +8,9 @@ use zyx_derive::Module;
 #[derive(Debug, Module)]
 pub struct Embedding {
     /// Vocabulary size
-    pub vocab_size: usize,
+    pub vocab_size: u64,
     /// Embedding size
-    pub embed_size: usize,
+    pub embed_size: u64,
     /// Weight
     pub weight: Tensor,
     /// Arange
@@ -19,7 +19,7 @@ pub struct Embedding {
 
 impl Embedding {
     /// new embedding layer
-    pub fn new(vocab_size: usize, embed_size: usize, dtype: DType) -> Result<Embedding, ZyxError> {
+    pub fn new(vocab_size: u64, embed_size: u64, dtype: DType) -> Result<Embedding, ZyxError> {
         Ok(Embedding {
             vocab_size,
             embed_size,
@@ -54,7 +54,7 @@ impl Embedding {
                 x_sh.iter()
                     .copied()
                     .chain([self.embed_size])
-                    .collect::<Vec<usize>>(),
+                    .collect::<Vec<u64>>(),
                 x.dtype(),
             ));
         }
@@ -66,14 +66,14 @@ impl Embedding {
                     .into(),
             ));
         }
-        let big_shp: Vec<usize> = x_sh
+        let big_shp: Vec<u64> = x_sh
             .iter()
             .copied()
             .chain([self.vocab_size, self.embed_size])
             .collect();
         let arange = self.arange.expand(big_shp.clone())?;
         let idx = x
-            .reshape(x_sh.into_iter().chain([1, 1]).collect::<Vec<usize>>())?
+            .reshape(x_sh.into_iter().chain([1, 1]).collect::<Vec<u64>>())?
             .expand(big_shp.clone())?;
         let vals = self.weight.expand(big_shp)?;
         (arange.equal(idx)?.cast(xdt) * vals).sum([2])
